+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## Version 3.2.0 (under dev)
 
+- Drop Python 2 support, Python 3.6 is now required.
 - Add support for Chinese full-stop.
 
 ## Version 3.1 (2020-03-07)